Add files via upload
Develop a C++ program that reads data from a text file named "input.txt" and stores it in an array of integers. The first line of the file contains an integer value that specifies the number of elements in the array, and the rest of the lines contain the actual elements. Then, your program should remove all duplicates from the array using a function named "remove_duplicates" that takes the array and its size as parameters. Finally, your program should output the deduplicated array, along with its length, to a text file named "output.txt".